fix: change price from BigDecimal -> Integer

+++ src/main/java/ru/shmelev/stomatologyapp/dto/RequestAppointmentCreate.java
@@ -36,7 +36,7 @@ public record RequestAppointmentCreate(
         String description,
 
         @NotBlank(message = "Введите стоимость приема ")
-        BigDecimal price
+        Integer price
 ) {
     public RequestAppointmentCreate() {
         this(null, null, null, null, null, null, null, null);
